refactor(#176): rename migrations files by template
